+++ pkg/urbit/vere/ward.c
@@ -64,49 +64,6 @@ u3_dire_free(u3_dire *dir_u)
   c3_free(dir_u);
 }
 
-/* u3_ovum_init: initialize an unlinked potential event
-*/
-u3_ovum*
-u3_ovum_init(c3_w     mil_w,
-             u3_noun    tar,
-             u3_noun    wir,
-             u3_noun    cad)
-{
-  u3_ovum* egg_u = c3_malloc(sizeof(*egg_u));
-  egg_u->car_u = 0;
-  egg_u->try_w = 0;
-  egg_u->ptr_v = 0;
-  egg_u->mil_w = mil_w;
-  egg_u->tar   = tar;
-  egg_u->wir   = wir;
-  egg_u->cad   = cad;
-
-  egg_u->pre_u = egg_u->nex_u = 0;
-
-  egg_u->cb_u.news_f = 0;
-  egg_u->cb_u.bail_f = 0;
-
-  //  spinner defaults
-  //
-  egg_u->pin_u.lab   = u3k(u3h(wir));
-  egg_u->pin_u.del_o = c3y;
-
-  return egg_u;
-}
-
-/* u3_ovum_free: dispose an unlinked potential event
-*/
-void
-u3_ovum_free(u3_ovum *egg_u)
-{
-  u3z(egg_u->pin_u.lab);
-  u3z(egg_u->tar);
-  u3z(egg_u->wir);
-  u3z(egg_u->cad);
-
-  c3_free(egg_u);
-}
-
 /* u3_pico_init(): initialize a scry request struct
 */
 u3_pico*
